// SWE645-HW3-AngularApp/src/app/models/student.model.ts
export class Student{
  studentId: string;
  firstName: string;
  lastName: string;
  phone: string;
  streetAddress: string;
  unit: string;
  zipcode: number;
  url: string;
  city: string;
  state: string;
  liked: string;
  interest: string;
  likelihood: string;
  graduationMonth: string;
  graduationYear: number;
  dateOfVisit: string;
  comments: string;
  email: string;

}